Resolving type mismatch in NN model evaluation

+++ du_stats/utils/nn_utils/nn_evaluation_util.py
@@ -8,7 +8,7 @@ from du_stats.utils.nn_utils.nn_training_util import accuracy_fn
 
 def nn_evaluation(model:nn.Module, X_test:np.array, y_test:np.array)->dict:
     try:
-        X_test, y_test = torch.from_numpy(X_test), torch.from_numpy(y_test)
+        X_test, y_test = torch.from_numpy(X_test).type(torch.float), torch.from_numpy(y_test).type(torch.long)
         model.eval()
         with torch.inference_mode():
             y_logits = model(X_test)
